learning modülü/requirements.txt
pytest
pytest-asyncio
pytest-cov
mock
pyjwt
aiohttp
pydub
nltk
validators
celery
redis
python-docx
PyMuPDF
SpeechRecognition
transformers